restructure/refactor curriculum page and login page
