--- @types/login.d.ts
@@ -1,8 +1,0 @@
-export type Values = {
-  username: string
-  password: string
-}
-
-export type Props = {
-  submitLogin: (values: Values) => void
-}
