+++ pages/login/types.d.ts
@@ -1,0 +1,4 @@
+export type Values = {
+  username: string
+  password: string
+}
